Moved MetaString to a new file
