+++ lib/CGameState.h
@@ -43,7 +43,7 @@ class CCreature;
 class CMap;
 struct StartInfo;
 struct SetObjectProperty;
-struct MetaString;
+class MetaString;
 struct CPack;
 class CSpell;
 struct TerrainTile;
