+++ lib/battle/Unit.cpp
@@ -13,6 +13,7 @@
 
 #include "../VCMI_Lib.h"
 #include "../CGeneralTextHandler.h"
+#include "../MetaString.h"
 #include "../NetPacksBase.h"
 
 #include "../serializer/JsonDeserializer.h"
